Flake 8 fixes
Implemented an improved read for the back-end fhir resources
Fixes the ability to request a single record while applying the patient filter.
Defaults the nav bar record generation to returning html with json

Wrote a document to  explain the use of the hash value for beneficiary matching to the back-end fhir platform.