Adds a string to autocomplete so the browser won't suggest users names for their application name (#757)
Change Details
After some research, I ended up only implementing one fix. The fix that you'll see is a little bit of a hack, but I have tested it and it works. Essentially there was this annoying behavior where when you go to register an application, the "Application Name" field would get autocomplete suggestions of what the user has entered in fields before - so I saw a lot of "John French" and so on.

Apparently, autcomplete="off" is deprecated and ignored by modern browsers. But if you put in a random string, the browser decides to not autocomplete. So again, it is a little bit hacky but it makes the experience nicer.

Things I researched and decided to leave as they are:

In Firefox, there is this annoying functionality where it constantly thinks the client id/secret are actually user credentials. So, when you go to edit an application sometimes, it prompts you to update your password. As I was reading about this - there ARE ways to get around this, but they seem overkill and kind of negate to benefits of a password field in general. But I think this was bothering me because I was constantly registering, deleting, and editing applications to test the Sandbox. In reality our sandbox users might have one or two applications and update them infrequently. So I think we should leave this as it is. Open for discussion, though.
Acceptance Validation
Looking for approval for the "name" input fix on the application registration.
Would love confirmation/discussion around the decision to ignore the Firefox password field behavior.
External References
BLUEBUTTON-1335
Security Implications
No security implications here at all.