BLUEBUTTON-1682 Reference SLS workflow docs (#773)
* Update mymedicare_cb app README

Link to confluence which contains the workflow, examples, and rationale
for the SLS / MyMedicare authentication workflow.

* Add README reference in code

As a dev I would want a link to relevant info as close as possible to
the code I'm writing. Rather than duplicating the Confluence link across
in the code (it's already in the README), reference the README in the
code.