Removing a magic URL
There was an invalid/old URL hardcoded as a string throughout the codebase. This has been replaced by references to the variable that is pulled in via the .env. (This was FHIR_SERVER["FHIR_URL"] from base.py.)

This is an interim commit to catch the initial changes; without this change, there were 40+ tests failing; now, there are 6, some of which may still be related.