Converting `Versions` to ints
We have many places in the code that assume the version number is an integer.

It would be nice to thoroughly refactor this out, but that isn't easy. As a result, I've changed the internal representation of

Versions.V1

from 'v1' to 1.

The reason for this is that we don't know all the places that the "version" gets used as an integer (e.g. placed in a request/response object as an integer), or a default value is set to an integer, or ... and... and...

So, for now, it is safer to use the symbolic representation of the version, but make sure that if the abstraction "leaks," that it leaks in a way that is consistent with legacy code.

There is more we could do here, and we might want to design this more intentionally. But, for now, this passess all tests (still/again), and it makes sure that we don't accidentally uncover a space where the version was expected to be an int, but it became a string.