Tests passing, waffle adjusted
The waffle switch is interesting/complex state. It does not reset with every application deploy; it is a set of values in the DB. As a result, changing the defaults does not change things (necessarily) as we do development or deploy.

Further, some constructs are *decorators*, and as a result, cannot be used in all contexts.

waffle_switch('...')(...)
                      ^
                      |
             This gets decorated

if we put the wrong "kind" of thing there, nothing happens, or worse, something breaks. (more often, nothing happens.) So, being aware of what constructs from the waffle library we're using, how we're using them, where we're using them, and whether or not they get evaluated at the point where the app is started (e.g. when defining URLs) or on every single request (eg. if we decorate a function or use various waffle_* constructs inside of functions) matters.