Returning None from get_application_from_data
This function feels like it should fail if we have no app given the identifiers available.

However, tests use this as part of their setup, meaning that if we fail, then the *setup* of unit tests fails.

This has been noted in other places. Changing the function to revert to the original behavior, and return `None` when no application can be found.