Stack talks to BFD
This worked for test. There is now more cleanup.

There's a notional split between "setting up to run" and "getting the variables into the container."

This moves all the work out of the container. It could, perhaps, all be moved into the container if the correct env was passed for secrets gathering?