Tooling for improved local dev experience (#1423)
* First commit of the devtools work.

This will configure and launch a stack. It isn't grabbing some things correctly, because of some more env-specific configuration that needs to be done. However, it's a solid start.

* Updating; looks like it builds/runs.

* Stack talks to BFD

This worked for test. There is now more cleanup.

There's a notional split between "setting up to run" and "getting the variables into the container."

This moves all the work out of the container. It could, perhaps, all be moved into the container if the correct env was passed for secrets gathering?

* A cleaner/quieter boot sequence.

* Improving local build.

* Fix where check happens

* adding extra around kion setup

* Add more checking, change params.

* Updated/all tests passing

* Removing script, out of scope

* Incremental/bug fixes.

* Fixes.

* Fixed sandbox FHIRv3 URL.

* Removing comments, unneeded prints.

---------

Co-authored-by: Brandon Wang <brandon.wang@icf.com>
Co-authored-by: James Demery <jamesdemery@navapbc.com>