Initial Django v5.2 migration spike
There are questions that want to be answered, but this should allow:

1. Standing up the stack on `master`
2. Taking it down
3. Doing a `make build-local` from this branch
4. Standing up the stack on this branch
5. Successfully launching

Therer was an error with the testclient; when we returned from auth, the `api_ver` seems not to be set. This implies any number of things, but the important first step was getting the app to stand up with a completely new set of modules.

Note that we use the `vendor` directory, under version control. This may not be strictly necessary in a containerized build world. This commit (currently) does not modify the tree when loading requirements.in and friends. To be discussed/revisited.