Removing TokenHasProtectedCap
After much sleuthing and conversation, we decided that trying to make the ProtectedCapabilities model work for C4DIC does not make sense. Further, HasDigitalInsuranceCardScope does *almost* the same checking; the only thing it does not check is the *method*, which is not a showstopper (or even a concern, really, because we already gate that elsewhere).