Local stack changes
The primary change is that we're not using the vendored files, but instead using the .txt directly.

Therefore, the container build now runs pip-install, and pulls down packages.

Apparently, I was not running `makemigrations` as part of the startup. That's an oversight. We should always makemigrations/migrate on every standup.