These are new, pinned .in files
This should hold us to where we were.

However, our build/deploy process has to change for this to be picked up.

That said, as a test, this would basically... change the build/deploy process, but no libraries. That would probably be the safest place to start.

Then, we can start bumping libraries towards a full Django 5.2 move.