Initial commit of ItsLog port to Django
This is a fast port.

Events are stored as <source,event> pairs, where the values are signed 64-bit integers. We hash string values into those integers, creating a (fast) 1-way path to storing small amounts of information while representing significantly more info to us, the end users. Also, this means we're storing an index (8 bytes), a timestamp (6-8 bytes) and two int64s (16 bytes) for every log message (48 bytes). That's roughly as small as it gets.

A dictionary is written as well; the dictionary maps <source,event> pairs back to plain text. This can be used when constructing visualizations.

The model is ~100 loc with comments; currently, I'm not collapsing/condensing the table, although that is part of the design intent... daily, we would compress the tens-to-hundreds of thousands of events down to fewer rows by counting things up ("summing," I guess).

This will be limited in performance.